refactor: introduce profit calculator
