+++ profit_calculator/profit_calculator.go
@@ -1,0 +1,5 @@
+package profit_calculator
+
+func CalculateProfit() {
+	
+}
